requirements.txt
openai
pandas
numpy
psycopg2
python-dotenv
langchain
langchain-openai
fastapi
uvicorn[standard]